Apparently not everything was pinned - it broke...
This leads me to think that Cargo.lock files can't always be trusted
when git repositories are involved. Maybe not everything was pinned
enough?

Weirdly enough I was unable to get it to work even after retroactively
putting the revs I thought were working.

Now it's in a state I have seen before. Even though it is trying to
build, some trait is not auto-implemented, probably because something
is out of sync.