Add templating support
This introduces a new `--template` option that can be
used to specify the path to a Tera template file.
This way, styling can be totally customized by the
user.
This required a better separation between the
"data aggregation" part and the "data representation"
part (the output) so some parts had to be moved around.