Add tokio `full` feature flag
As of tokio 0.2, features are hidden behind flags.
For example, macros require the `macro` feature.
See https://github.com/tokio-rs/tokio/issues/1833
for a more detailed description.

The easiest way to stay up-to-date is by enabling
the `full` feature flag, which is analog to the
previous behavior. This fixes `cargo install`.